% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasisGraphicalLasso.R
\docType{package}
\name{BasisGraphicalLasso}
\alias{BasisGraphicalLasso}
\title{Basis Graphical Lasso}
\description{
Estimates a (sparse) precision matrix for the coefficients of the basis functions in the basis graphical lasso model.
}
