% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nugget_estimate.R
\name{nugget_estimate}
\alias{nugget_estimate}
\title{Nugget Estimate}
\usage{
nugget_estimate(Phi_Phi, Phi_S_Phi, trS, n)
}
\arguments{
\item{Phi_Phi}{Inner product of basis matrices, \eqn{\Phi'\Phi}. Computed as \code{crossprod(Phi)}, where \code{Phi} has n rows corresponding to locations and l columns corresponding to the basis functions evaluated at those locations.}

\item{Phi_S_Phi}{Inner product of the basis matrices and data, \eqn{\Phi'S\Phi}. This is where the data directly enters the algorithm. Note: do not compute sample covariance S explicitly. With \code{dat} as the n by m matrix with columns corresponding to realizations of the mean zero spatial field, we have \eqn{S=XX'/m}. So we can compute as \eqn{\Phi'S\Phi} as \code{tcrossprod(crossprod(Phi, dat))/m}.}

\item{trS}{Trace of empirical covariance matrix.}

\item{n}{Number of locations.}
}
\value{
Estimate of the nugget variance.
}
\description{
Estimates the nugget variance \eqn{\tau^2} under the simplified assumption that \eqn{Q = \alpha I}. If not desired, user can initially pass \code{tau_sq}.
}
\details{
Performs a joint optimization over \eqn{\alpha,\tau^2} in the -2*negative log likelihood with respect to \eqn{Q = \alpha I} and \eqn{\tau^2}. Calls L-BFBS-B using \code{optim}.
}
\examples{
basis.setup <- BGLBasisSetup(y=tmin$data,locs=tmin$lon.lat.proj,basis="LatticeKrig",
     crossvalidation=FALSE, NC=30,nlevel=1)
Phi_Phi <- basis.setup$Phi_Phi
Phi_S_Phi <- basis.setup$Phi_S_Phi
trS <- basis.setup$trS
tau_sq <- nugget_estimate(Phi_Phi,Phi_S_Phi,trS,n=dim(tmin$lon.lat)[1])
}
