% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penaltymatrixsetup.R
\name{penaltymatrixsetup}
\alias{penaltymatrixsetup}
\title{Penalty Matrix Setup}
\usage{
penaltymatrixsetup(
  lambda,
  zero.diagonal.penalty = TRUE,
  l,
  basisdistancematrix = NULL
)
}
\arguments{
\item{lambda}{Scalar or matrix.}

\item{zero.diagonal.penalty}{Zero penalty along the diagonals of precision. Default: TRUE.}

\item{l}{Number of basis functions.}

\item{basisdistancematrix}{The distance between LatticeKrig basis centers. Default: NULL.}
}
\value{
Penalty matrix.
}
\description{
Creates the penalty matrix.
}
\details{
Returns a penalty matrix.
}
