% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpenalized_neglikelihood_Q.R
\name{unpenalized_neglikelihood_Q}
\alias{unpenalized_neglikelihood_Q}
\title{Evaluate Negative Log Likelihood}
\usage{
unpenalized_neglikelihood_Q(Q, Phi_Dinv_Phi, Phi_Dinv_S_Dinv_Phi)
}
\arguments{
\item{Q}{Precision matrix.}

\item{Phi_Dinv_Phi}{Inner product of basis matrices, \eqn{\Phi'D^{-1}\Phi}. Typically obtained from using \code{BGLBasisSetup.R} with \eqn{D = \tau^2 I}.}

\item{Phi_Dinv_S_Dinv_Phi}{Inner product of the basis matrices and data, \eqn{\Phi'D^{-1}SD^{-1}\Phi}. Typically obtained from using \code{BGLBasisSetup} with \eqn{D = \tau^2 I}. This is where the data directly enters the algorithm. Note: do not compute sample covariance S explicitly. With \code{dat} as the n by m matrix with columns corresponding to realizations of the mean zero spatial field, we have \eqn{S=XX'/m}. So, setting D = I for simplicity, we can compute as \eqn{\Phi'S\Phi} as \code{tcrossprod(crossprod(Phi, dat))/m}.}
}
\value{
-2*log likelihood, ignoring additive constants not depending on Q.
}
\description{
Evaluates the negative log likelihood without the normalizing constant.
}
\details{
This returns negative two times the log likelihood, ignoring anything that doesn't depend on Q or \eqn{\tau^2}. Note that the evaluation only requires matrix computations in the number of the basis functions since \eqn{Phi'Phi} and \eqn{Phi'SPhi} are already computed.
}
